// Front-Billing/microservicio/projects/billingclient/src/services/swaggerBillingAPI/model/invoiceResponse.ts
/**
 * Billing API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

/**
 * Model represent a Invoice on database
 */
export interface InvoiceResponse { 
    /**
     * Unique Id of invoice
     */
    invoiceId: number;
    /**
     * Unique Id of customer that represent the owner of invoice
     */
    customer: number;
    /**
     * Number given in fisical invoice
     */
    number: string;
    detail?: string;
    amount?: number;
}